synthesizable FPGA benchmark

+++ modules/airblue/lib/HostControl/Environment/check_ber.h
@@ -3,7 +3,7 @@
 
 int get_rate();
 int get_packet_size();
-unsigned int get_finish_cycles();
-int check_ber(int errors, int total);
+long long get_finish_cycles();
+long long check_ber(long long errors, long long total);
 
 #endif
